+++ src/shared/ui/icon/index.ts
@@ -1,2 +1,5 @@
 export * from "./ui";
 export * as iconLib from "./lib";
+
+// временный костыль что бы импорты не сломались
+export * from "./lib";
